rename package to inmem
